requirements.txt
fastapi[standard] # Include all required dependencies
pydantic==2.11.4
python-dotenv==1.1.0
starlette==0.46.2
uvicorn==0.34.2
pydantic-settings==2.9.1
requests==2.32.4
opencv-python==4.11.0.86
pyserial
pillow==11.2.1
torch
python-dotenv==1.1.0
python-jose==3.5.0
passlib==1.7.4
torchvision
psycopg2-binary
sqlalchemy
# Code quality
# ------------------------------------------------------------------------------
black==25.1.0
pre-commit==3.7.0
pytest==8.3.5